Avoid requiring some node dependencies using the package.json browser field
Change-type: patch
