+++ lib/progress.ts
@@ -156,35 +156,11 @@ export function estimate(
 		);
 
 		if (!isBrowser && utils.isResponseCompressed(response)) {
-			// eslint-disable-next-line @typescript-eslint/no-var-requires
-			const zlib = require('zlib') as typeof import('zlib');
+			const { createGunzip } =
+				// eslint-disable-next-line @typescript-eslint/no-var-requires
+				require('./conditional-imports') as typeof import('./conditional-imports');
 
-			// Be more lenient with decoding compressed responses, since (very rarely)
-			// servers send slightly invalid gzip responses that are still accepted
-			// by common browsers.
-			// Always using Z_SYNC_FLUSH is what cURL does.
-			let zlibOptions = {
-				flush: zlib.constants.Z_SYNC_FLUSH,
-				finishFlush: zlib.constants.Z_SYNC_FLUSH,
-			};
-
-			// Allow overriding this behaviour by setting the ZLIB_FLUSH env var
-			// to one of the allowed zlib flush values (process.env.ZLIB_FLUSH="Z_NO_FLUSH").
-			// https://github.com/nodejs/node/blob/master/doc/api/zlib.md#zlib-constants
-			if (process.env.ZLIB_FLUSH && process.env.ZLIB_FLUSH in zlib.constants) {
-				zlibOptions = {
-					flush:
-						zlib.constants[
-							process.env.ZLIB_FLUSH as keyof typeof zlib.constants
-						],
-					finishFlush:
-						zlib.constants[
-							process.env.ZLIB_FLUSH as keyof typeof zlib.constants
-						],
-				};
-			}
-
-			const gunzip = zlib.createGunzip(zlibOptions);
+			const gunzip = createGunzip();
 			gunzip.on('error', (e) => output.emit('error', e));
 
 			// Uncompress after or before piping through progress
